管线和管线集合数据结构改变

+++ app/src/main/java/com/barisetech/www/workmanage/bean/pipe/PipeInfo.java
@@ -63,7 +63,7 @@ public class PipeInfo implements Serializable {
 
     public int PipeId;
     public String Name;
-    public int Length;
+    public double Length;
     public String Company;
     public String PipeMaterial;
     public int Speed;
